requirements.txt
Flask
mysql-connector-python
python-dotenv
aiohttp
schedule
python-dateutil
pyppeteer
pytz
fake_useragent